requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python
pillow
scikit-image
matplotlib
tqdm
google-generativeai
python-dotenv
reportlab
